refactor(@nestjs/graphql): rename helper
- rename `MixinType` into `IntersectionType`
- use intersection type inside `Type` rather than out
