+++ lib/type-helpers/index.ts
@@ -1,4 +1,4 @@
-export * from './mixin-type.helper';
+export * from './intersection-type.helper';
 export * from './omit-type.helper';
 export * from './partial-type.helper';
 export * from './pick-type.helper';
